requirements.txt
tensorboardX
pytorch
torchtext
tqdm
